Add event participation feature with backend and frontend support
Implemented event participation management, including backend APIs, repository, and DTOs. Updated frontend to display participant statuses and allow status changes. Database and migration changes were added to support the new feature.

+++ icy_backend/src/main/resources/db/migration/V3__add_brand.sql
@@ -6,7 +6,7 @@ CREATE TABLE brand (
 );
 
 -- Ajout de la colonne brand_id à la table ships
-ALTER TABLE ships ADD COLUMN brand_id INT;
+ALTER TABLE ships ADD COLUMN brand_id BIGINT;
 
 -- Création de la contrainte FK entre ships et brand
 ALTER TABLE ships
